# frontend/openapi/context69.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "context69",
    "description": "",
    "license": {
      "name": "Apache-2.0",
      "identifier": "Apache-2.0"
    },
    "version": "0.8.1"
  },
  "paths": {
...
    "/v1/library/files/upload": {
      "post": {
        "tags": [],
        "operationId": "upload_library_files",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "string"
              }
            }
          },
          "required": true
        },
        "responses": {
          "400": {
            "description": "Invalid upload",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/library/files/{file_id}": {
      "get": {
        "tags": [],
        "operationId": "get_library_file",
        "parameters": [
          {
            "name": "file_id",
            "in": "path",
            "description": "File id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Library file details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LibraryFileDetailResponse"
                }
              }
            }
          },
          "404": {
            "description": "File not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [],
        "operationId": "delete_library_file",
        "parameters": [
          {
            "name": "file_id",
            "in": "path",
            "description": "File id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "File not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/library/files/{file_id}/move": {
      "post": {
        "tags": [],
        "operationId": "move_library_file",
        "parameters": [
          {
            "name": "file_id",
            "in": "path",
            "description": "File id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MoveFileRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Moved library file",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LibraryFileDetailResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid move request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "File not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/library/folders": {
      "post": {
        "tags": [],
        "operationId": "create_library_folder",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFolderRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created folder",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LibraryFolderResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid folder request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/library/folders/{folder_id}": {
      "delete": {
        "tags": [],
        "operationId": "delete_library_folder",
        "parameters": [
          {
            "name": "folder_id",
            "in": "path",
            "description": "Folder id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Folder delete task accepted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskRef"
                }
              }
            }
          },
          "404": {
            "description": "Folder not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/library/folders/{folder_id}/move": {
      "post": {
        "tags": [],
        "operationId": "move_library_folder",
        "parameters": [
          {
            "name": "folder_id",
            "in": "path",
            "description": "Folder id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MoveFolderRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Moved folder",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LibraryFolderResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid move request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Folder not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/library/resources": {
      "get": {
        "tags": [],
        "operationId": "get_library_resources",
        "parameters": [
          {
            "name": "folder_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          },
          {
            "name": "query",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/LibraryIngestStatus"
            }
          },
          {
            "name": "sort_by",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/LibraryResourceSortBy"
            }
          },
          {
            "name": "sort_direction",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/SortDirection"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated library resources",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LibraryResourcePageResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid pagination parameters"
          },
          "404": {
            "description": "Folder not found"
          }
        }
      }
    },
    "/v1/library/texts": {
      "post": {
        "tags": [],
        "operationId": "create_library_text",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTextRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "400": {
            "description": "Invalid text payload",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/library/tree": {
      "get": {
        "tags": [],
        "operationId": "get_library_tree",
        "responses": {
          "200": {
            "description": "Library directory tree",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LibraryTreeResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiErrorResponse": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
...
      "CreateFolderRequest": {
        "type": "object",
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "parent_folder_id": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          }
        }
      },
...
      "CreateTextRequest": {
        "type": "object",
        "required": [
          "title",
          "content"
        ],
        "properties": {
          "content": {
            "type": "string"
          },
          "content_format": {
            "$ref": "#/components/schemas/LibraryTextContentFormat"
          },
          "folder_id": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          },
          "source_uri": {
            "type": [
              "string",
              "null"
            ]
          },
          "summary": {
            "type": [
              "string",
              "null"
            ]
          },
          "title": {
            "type": "string"
          },
          "translation": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/TranslationDirective"
              }
            ]
          }
        }
      },
...
      "LibraryDocumentSectionPreview": {
        "type": "object",
        "required": [
          "document_id",
          "section_key",
          "section_label",
          "sort_order",
          "title",
          "preview_text"
        ],
        "properties": {
          "content_format": {
            "$ref": "#/components/schemas/LibraryPreviewContentFormat"
          },
          "document_id": {
            "type": "integer",
            "format": "int64"
          },
          "preview_text": {
            "type": "string"
          },
          "section_key": {
            "type": "string"
          },
          "section_label": {
            "type": "string"
          },
          "sort_order": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "type": "string"
          }
        }
      },
      "LibraryFileDetailResponse": {
        "type": "object",
        "required": [
          "file_id",
          "group_key",
          "group_path",
          "visibility",
          "folder_path",
          "filename",
          "media_type",
          "size_bytes",
          "sha256",
          "ingest_status",
          "created_at",
          "updated_at",
          "sections"
        ],
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "error_message": {
            "type": [
              "string",
              "null"
            ]
          },
          "file_id": {
            "type": "string",
            "format": "uuid"
          },
          "filename": {
            "type": "string"
          },
          "folder_id": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          },
          "folder_path": {
            "type": "string"
          },
          "group_key": {
            "type": "string"
          },
          "group_path": {
            "type": "string"
          },
          "ingest_status": {
            "$ref": "#/components/schemas/LibraryIngestStatus"
          },
          "ingested_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "media_type": {
            "type": "string"
          },
          "sections": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LibraryDocumentSectionPreview"
            }
          },
          "sha256": {
            "type": "string"
          },
          "size_bytes": {
            "type": "integer",
            "format": "int64"
          },
          "source_available": {
            "type": "boolean"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          },
          "visibility": {
            "$ref": "#/components/schemas/Visibility"
          }
        }
      },
...
      "LibraryFolderResponse": {
        "type": "object",
        "required": [
          "folder_id",
          "group_key",
          "group_path",
          "visibility",
          "name",
          "path",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "folder_id": {
            "type": "string",
            "format": "uuid"
          },
          "group_key": {
            "type": "string"
          },
          "group_path": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "parent_folder_id": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          },
          "path": {
            "type": "string"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          },
          "visibility": {
            "$ref": "#/components/schemas/Visibility"
          }
        }
      },
...
      "LibraryIngestStatus": {
        "type": "string",
        "enum": [
          "pending",
          "running",
          "succeeded",
          "failed"
        ]
      },
...
      "LibraryResourceItem": {
        "type": "object",
        "required": [
          "kind",
          "id",
          "group_key",
          "group_path",
          "visibility",
          "name",
          "child_folder_count",
          "file_count",
          "processing_count",
          "is_source_folder",
          "is_source_records_folder",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "child_folder_count": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "error_message": {
            "type": [
              "string",
              "null"
            ]
          },
          "file_count": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "group_key": {
            "type": "string"
          },
          "group_path": {
            "type": "string"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "ingest_status": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/LibraryIngestStatus"
              }
            ]
          },
          "is_source_folder": {
            "type": "boolean"
          },
          "is_source_records_folder": {
            "type": "boolean"
          },
          "kind": {
            "$ref": "#/components/schemas/LibraryResourceKind"
          },
          "media_type": {
            "type": [
              "string",
              "null"
            ]
          },
          "name": {
            "type": "string"
          },
          "parent_folder_id": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          },
          "processing_count": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "size_bytes": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          },
          "visibility": {
            "$ref": "#/components/schemas/Visibility"
          }
        }
      },
...
      "LibraryResourcePageResponse": {
        "type": "object",
        "required": [
          "items",
          "pagination"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LibraryResourceItem"
            }
          },
          "pagination": {
            "$ref": "#/components/schemas/Pagination"
          }
        }
      },
      "LibraryResourceSortBy": {
        "type": "string",
        "enum": [
          "name",
          "type",
          "status",
          "size",
          "updated_at"
        ]
      },
...
      "LibraryTreeResponse": {
        "type": "object",
        "required": [
          "root"
        ],
        "properties": {
          "root": {
            "$ref": "#/components/schemas/LibraryFolderNode"
          }
        }
      },
...
      "MoveFileRequest": {
        "type": "object",
        "properties": {
          "target_folder_id": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          }
        }
      },
      "MoveFolderRequest": {
        "type": "object",
        "properties": {
          "target_folder_id": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          }
        }
      },
...
      "Pagination": {
        "type": "object",
        "required": [
          "page",
          "page_size",
          "total",
          "total_pages"
        ],
        "properties": {
          "page": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "page_size": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "total": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "total_pages": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          }
        }
      },
...
      "SortDirection": {
        "type": "string",
        "enum": [
          "asc",
          "desc"
        ]
      },
...
      "TaskRef": {
        "type": "object",
        "required": [
          "task_id",
          "item_ids"
        ],
        "properties": {
          "task_id": {
            "type": "string",
            "format": "uuid"
          },
          "item_ids": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          }
        }
      },
...
      "Visibility": {
        "type": "string",
        "enum": [
          "public",
          "private"
        ]
      }
    }
  },
...
}